Encrypt Credit Card Number
The cart implementation now has a method to transmit the credit card
number directly to UltraCart.   This prevents it from going through your
rest_proxy.php script.  This is a big security win.   With the old
method, your web site might be subject to additional requirements to be
PCI Compliant.  With the credit card going  directly to UltraCart, you
don't have to worry (as much) about a site comprimise.

Look for the javascript method app.commonFunctions.storeCreditCard for
details.